+++ connect-react/src/Pages/Dashboard/Dashboard.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Row from 'react-bootstrap/Row';
-import Container from "../components/Container";
-import SideMenu from "../components/SideMenu";
+import Container from "./Container";
+import SideMenu from "./SideMenu";
 
 const Dashboard = () => {
 	return (
