+++ connect-react/src/Pages/Dashboard/SideMenu.js
