+++ connect-react/src/Pages/Dashboard/SubPages/Feed/FeedCard.js
