+++ connect-react/src/Pages/Dashboard/SubPages/Venues/VenueFigure.js
