+++ connect-react/src/Pages/Dashboard/SubPages/Venues/VenuePage.js
@@ -3,7 +3,7 @@ import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Figure from 'react-bootstrap/Figure';
-import Mount_Snow from '../images/mount_snow_bg.png'
+import Mount_Snow from '../../../../images/mount_snow_bg.png'
 //import VenueButtonMenu from './VenueButtonMenu';
 import Button from 'react-bootstrap/Button';
 import ButtonGroup from 'react-bootstrap/ButtonGroup';
